refactor(translateMagicBlocks): split into lib/translate/ folder
Splits the single-file translateMagicBlocks helper into a four-file
module under lib/translate/ with a per-type translator registry, per PR
review feedback from @eaglethrost and @maximilianfalco on #1473.

Adds JSDoc on the public entry and the image translator, plus a
comment noting MAGIC_BLOCK_OPEN_RE is a prefix subset of MAGIC_BLOCK_REGEX
in lib/utils/extractMagicBlocks.ts. Adding a new block-type translator
is now one new file plus one line in the translators map.

No behavior change for the image case; serializer (mdxishMdastToMd +
extended toAttributes) is unchanged.
