+++ lib/index.ts
@@ -19,4 +19,4 @@ export { default as run } from './run';
 export { default as tags } from './tags';
 export { default as mdxishTags } from './mdxishTags';
 export { default as stripComments } from './stripComments';
-export { default as translateMagicBlocks } from './translateMagicBlocks';
+export { default as translateMagicBlocks } from './translate';
