+++ lib/translate/image.ts
@@ -1,51 +1,18 @@
-import type { MdastNode, MagicBlockImage } from '../processor/transform/mdxish/magic-blocks/types';
+import type { MdastNode } from '../../processor/transform/mdxish/magic-blocks/types';
 import type { Root as MdastRoot } from 'mdast';
 import type { MdxJsxFlowElement } from 'mdast-util-mdx-jsx';
 
-import remarkParse from 'remark-parse';
-import { unified } from 'unified';
+import { toAttributes } from '../../processor/utils';
+import { mdxishMdastToMd } from '../mdxish';
 
-import magicBlockTransformer from '../processor/transform/mdxish/magic-blocks/magic-block-transformer';
-import { toAttributes } from '../processor/utils';
+import {
+  getChildren,
+  isRecord,
+  type MagicBlockFigure,
+  type MagicBlockImageNode,
+} from './utils';
 
-import { magicBlockFromMarkdown } from './mdast-util/magic-block';
-import { mdxishMdastToMd } from './mdxish';
-import { magicBlock } from './micromark/magic-block';
-import { MAGIC_BLOCK_REGEX } from './utils/extractMagicBlocks';
-
-type MagicBlockTranslator = (raw: string) => string;
-
-interface MagicBlockFigure extends MdastNode {
-  children: MdastNode[];
-  type: 'figure';
-}
-
-type MagicBlockImageNode = MagicBlockImage & MdastNode;
-
-const MAGIC_BLOCK_OPEN_RE = /^\[block:([^\]]{1,100})\]/;
 const IMAGE_BLOCK_BODY_RE = /^\[block:image\]([\s\S]*)\[\/block\]$/;
-
-const processor = unified()
-  .data('micromarkExtensions', [magicBlock()])
-  .data('fromMarkdownExtensions', [magicBlockFromMarkdown()])
-  .use(remarkParse)
-  .use(magicBlockTransformer, { safeMode: true });
-
-function countNewlines(value: string) {
-  return value.split('\n').length - 1;
-}
-
-function isRecord(value: unknown): value is Record<string, unknown> {
-  return typeof value === 'object' && value !== null && !Array.isArray(value);
-}
-
-function isMdastNode(value: unknown): value is MdastNode {
-  return isRecord(value) && typeof value.type === 'string';
-}
-
-function isMdastRoot(value: unknown): value is MdastRoot {
-  return isRecord(value) && value.type === 'root' && Array.isArray(value.children);
-}
 
 function isMagicBlockImage(node: MdastNode): node is MagicBlockImageNode {
   return node.type === 'image';
@@ -53,10 +20,6 @@ function isMagicBlockImage(node: MdastNode): node is MagicBlockImageNode {
 
 function isMagicBlockFigure(node: MdastNode): node is MagicBlockFigure {
   return node.type === 'figure' && Array.isArray(node.children);
-}
-
-function getChildren(value: MdastNode | MdastRoot) {
-  return Array.isArray(value.children) ? value.children.filter(isMdastNode) : [];
 }
 
 function hasImageBlockData(raw: string) {
@@ -85,6 +48,11 @@ function stringifyCaption(node: MdastNode): string {
 
 function imageToMdx(node: MagicBlockImageNode, caption?: string) {
   const hProperties = node.data?.hProperties ?? {};
+  // hProperties.border declared as `string` upstream; runtime is actually boolean
+  // (see magic-block-transformer.ts:424). Read as unknown and coerce here rather
+  // than widen the upstream type. Drift cleanup is filed as a follow-up.
+  const rawBorder = (hProperties as { border?: unknown }).border;
+  const borderValue = typeof rawBorder === 'boolean' ? rawBorder : undefined;
   const src = node.url;
 
   if (!src) return null;
@@ -92,7 +60,7 @@ function imageToMdx(node: MagicBlockImageNode, caption?: string) {
   const attributes = {
     align: hProperties.align,
     alt: node.alt ?? '',
-    border: hProperties.border,
+    border: borderValue,
     caption,
     title: node.title || undefined,
     width: hProperties.width,
@@ -140,42 +108,16 @@ function stringifyMdxImage(tree: MdastRoot) {
   return mdxishMdastToMd({ type: 'root', children: [mdxNode] }).trim();
 }
 
-function translateImageBlock(raw: string) {
+/**
+ * Translates a single [block:image] span into <Image /> JSX via the magic-block
+ * transformer + mdxishMdastToMd. Returns the original span on parse failure or
+ * when no figure/image node is produced. Newline padding is applied by the caller.
+ */
+export default function translateImageBlock(raw: string, tree: MdastRoot) {
   if (!hasImageBlockData(raw)) return raw;
 
-  try {
-    const tree = processor.runSync(processor.parse(raw));
-    if (!isMdastRoot(tree)) return raw;
+  const translated = stringifyMdxImage(tree);
+  if (!translated) return raw;
 
-    const translated = stringifyMdxImage(tree);
-    if (!translated) return raw;
-
-    const originalNewlineCount = countNewlines(raw);
-    const translatedNewlineCount = countNewlines(translated);
-    if (translatedNewlineCount > originalNewlineCount) return raw;
-
-    return translated + '\n'.repeat(originalNewlineCount - translatedNewlineCount);
-  } catch {
-    return raw;
-  }
-}
-
-const translators: Partial<Record<string, MagicBlockTranslator>> = {
-  image: translateImageBlock,
-};
-
-function translateMagicBlock(raw: string) {
-  const blockType = raw.match(MAGIC_BLOCK_OPEN_RE)?.[1];
-  const translator = blockType ? translators[blockType] : undefined;
-
-  return translator ? translator(raw) : raw;
-}
-
-/**
- * Translates supported legacy magic blocks into MDX-shaped markdown while
- * preserving the source document's line count.
- */
-export default function translateMagicBlocks(content: string) {
-  MAGIC_BLOCK_REGEX.lastIndex = 0;
-  return content.replace(MAGIC_BLOCK_REGEX, match => translateMagicBlock(match));
+  return translated;
 }
