+++ lib/translate/index.ts
@@ -1,0 +1,3 @@
+import translateMagicBlocks from './translateMagicBlocks';
+
+export default translateMagicBlocks;
